ViewModelファイル追加
