+++ Kei3MapAppPackage/Sources/Feature/CreateAccount/SetAccountThumbnailView.swift
@@ -1,32 +1,4 @@
 import SwiftUI
-import Repository
-import _PhotosUI_SwiftUI
-
-@MainActor
-public class SetAccountThumbnailViewModel: ObservableObject {
-  let accountName: String
-  let onCreated: () -> Void
-
-  @Published var photoItem: PhotosPickerItem? {
-    didSet {
-      setProfileUIImage()
-    }
-  }
-  @Published var profileImage: UIImage?
-  @Published var isImagePickerPresented = false
-
-  public init(accountName: String, onCreated: @escaping () -> Void) {
-    self.accountName = accountName
-    self.onCreated = onCreated
-  }
-
-  // UIImageに変換
-  func setProfileUIImage() {
-    Task {
-      profileImage = await photoItem?.toUIImage()
-    }
-  }
-}
 
 public struct SetAccountThumbnailView: View {
   @StateObject private var viewModel: SetAccountThumbnailViewModel
@@ -58,7 +30,6 @@ public struct SetAccountThumbnailView: View {
           let viewModel = ConfirmCreateAccountViewModel(
             accountName: viewModel.accountName,
             profileImage: viewModel.profileImage,
-            accountManageRepository: AccountManageRepository(),
             onCreated: viewModel.onCreated)
           ConfirmCreateAccountView(viewModel: viewModel)
         } label: {
