+++ Kei3MapAppPackage/Sources/Feature/Post/PostDetailView.swift
@@ -2,20 +2,11 @@ import Entity
 import SwiftUI
 import Kingfisher
 
-@MainActor
-public class PostDetailViewModel: ObservableObject {
+public struct PostDetailView: View {
   let postItem: Post
 
   public init(postItem: Post) {
     self.postItem = postItem
-  }
-}
-
-public struct PostDetailView: View {
-  @StateObject private var viewModel: PostDetailViewModel
-
-  public init(postItem: Post) {
-    _viewModel = StateObject(wrappedValue: PostDetailViewModel(postItem: postItem))
   }
 
   public var body: some View {
@@ -23,17 +14,17 @@ public struct PostDetailView: View {
       RoundedRectangle(cornerRadius: 0)
         .aspectRatio(1, contentMode: .fit)
         .overlay {
-          KFImage(viewModel.postItem.postImageURL)
+          KFImage(postItem.postImageURL)
             .resizable()
             .scaledToFill()
         }
         .clipShape(RoundedRectangle(cornerRadius: 4))
         .padding(.top, 16)
-      Text(viewModel.postItem.addressString)
+      Text(postItem.addressString)
         .foregroundStyle(.gray)
         .frame(maxWidth: .infinity, alignment: .leading)
         .padding(.horizontal, 16)
-      Text(viewModel.postItem.postText)
+      Text(postItem.postText)
         .frame(maxWidth: .infinity, alignment: .leading)
         .padding(.horizontal, 16)
         .padding(.top, 16)
