--- Kei3MapAppPackage/Sources/Feature/Root/MainTabView.swift
@@ -1,7 +1,0 @@
-import SwiftUI
-
-struct MainTabView: View {
-  var body: some View {
-    Text("TabView")
-  }
-}
